+++ NEWS.md
@@ -3,6 +3,7 @@ profoc v0.5.X (Release date: 2021-XX-XX)
 
 Changes:
 
+* `loss_function` lets you now specify "quantile", "expectile" or "percentage". All functions are generalized as in [Gneitling 2009](https://arxiv.org/abs/0912.0902). The power can be scaled by `loss_parameter`. The latter defaults to 1 wich leads to the well known quantile, sqaurred and absolute percentage loss.
 * Add `lead_time` parameter. offset for expert forecasts. Defaults to 0, which means that experts predict t+1 at t. Setting this to h means experts predictions refer to t+1+h at time t. The weight updates delay accordingly.
 * `forget_performance` was added. It defines which share of the past performance will not be considered for selecting the best parameter combination.
 * The `forget` parameter was renamed to `forget_regret` to underline its reference to the regret.